requirements.txt
# Flask dependencies
flask
flask-cors

# Core dependencies
requests
python-dotenv

# OpenAI integration
openai
pydantic

# HTML processing
html2text

# Timezone handling
pytz

# SSL certificate generation
pyopenssl

# Optional: Message file parsing (for .msg attachments)
extract-msg

# Fuzzy string matching for vendor lookup
fuzzywuzzy
python-Levenshtein